# api-gateway/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Auth, Role-management, Product, User",
        "title": "THIRD EXAM",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:7070",
    "paths": {
...
        "/v1/rbac/add/policy": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Add policy to a role",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Role-management"
                ],
                "summary": "add policy to a role",
                "parameters": [
                    {
                        "type": "string",
                        "description": "super-username",
                        "name": "super-username",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "super-password",
                        "name": "super-password",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "policy",
                        "name": "policy",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.AddPolicyRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.SuperAdminMessage"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/v1/rbac/delete/policy": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete policy",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Role-management"
                ],
                "summary": "delete policy",
                "parameters": [
                    {
                        "type": "string",
                        "description": "super-username",
                        "name": "super-username",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "super-password",
                        "name": "super-password",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "policy",
                        "name": "policy",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.AddPolicyRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.SuperAdminMessage"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/v1/rbac/policies/{role}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get all policies of a role",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Role-management"
                ],
                "summary": "get all policies of a role",
                "parameters": [
                    {
                        "type": "string",
                        "description": "super-username",
                        "name": "super-username",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "super-password",
                        "name": "super-password",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "role",
                        "name": "role",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.ListRolePolicyResp"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/v1/rbac/roles": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get all roles",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Role-management"
                ],
                "summary": "get all roles",
                "parameters": [
                    {
                        "type": "string",
                        "description": "super-username",
                        "name": "super-username",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "super-password",
                        "name": "super-password",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.RbacAllRolesResp"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.AddPolicyRequest": {
            "type": "object",
            "properties": {
                "policy": {
                    "$ref": "#/definitions/models.Policy"
                }
            }
        },
...
        "models.ListRolePolicyResp": {
            "type": "object",
            "properties": {
                "policies": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Policy"
                    }
                }
            }
        },
...
        "models.Policy": {
            "type": "object",
            "properties": {
                "endpoint": {
                    "type": "string"
                },
                "method": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                }
            }
        },
...
        "models.RbacAllRolesResp": {
            "type": "object",
            "properties": {
                "roles": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "models.SuperAdminMessage": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}